+++ sdk/resourcemanager/devhub/armdevhub/models.go
@@ -26,6 +26,48 @@ type ACR struct {
 	AcrSubscriptionID *string `json:"acrSubscriptionId,omitempty"`
 }
 
+// ArtifactGenerationProperties - Properties used for generating artifacts such as Dockerfiles and manifests.
+type ArtifactGenerationProperties struct {
+	// The name of the app.
+	AppName *string `json:"appName,omitempty"`
+
+	// The version of the language image used for building the code in the generated dockerfile.
+	BuilderVersion *string `json:"builderVersion,omitempty"`
+
+	// The mode of generation to be used for generating Dockerfiles.
+	DockerfileGenerationMode *DockerfileGenerationMode `json:"dockerfileGenerationMode,omitempty"`
+
+	// The directory to output the generated Dockerfile to.
+	DockerfileOutputDirectory *string `json:"dockerfileOutputDirectory,omitempty"`
+
+	// The programming language used.
+	GenerationLanguage *GenerationLanguage `json:"generationLanguage,omitempty"`
+
+	// The name of the image to be generated.
+	ImageName *string `json:"imageName,omitempty"`
+
+	// The tag to apply to the generated image.
+	ImageTag *string `json:"imageTag,omitempty"`
+
+	// The version of the language image used for execution in the generated dockerfile.
+	LanguageVersion *string `json:"languageVersion,omitempty"`
+
+	// The mode of generation to be used for generating Manifest.
+	ManifestGenerationMode *ManifestGenerationMode `json:"manifestGenerationMode,omitempty"`
+
+	// The directory to output the generated manifests to.
+	ManifestOutputDirectory *string `json:"manifestOutputDirectory,omitempty"`
+
+	// Determines the type of manifests to be generated.
+	ManifestType *GenerationManifestType `json:"manifestType,omitempty"`
+
+	// The namespace to deploy the application to.
+	Namespace *string `json:"namespace,omitempty"`
+
+	// The port the application is exposed on.
+	Port *string `json:"port,omitempty"`
+}
+
 // DeleteWorkflowResponse - delete response if content must be provided on delete operation
 type DeleteWorkflowResponse struct {
 	// delete status message
@@ -47,6 +89,12 @@ type DeploymentProperties struct {
 	Overrides map[string]*string `json:"overrides,omitempty"`
 }
 
+// DeveloperHubServiceClientGeneratePreviewArtifactsOptions contains the optional parameters for the DeveloperHubServiceClient.GeneratePreviewArtifacts
+// method.
+type DeveloperHubServiceClientGeneratePreviewArtifactsOptions struct {
+	Parameters *ArtifactGenerationProperties
+}
+
 // DeveloperHubServiceClientGitHubOAuthCallbackOptions contains the optional parameters for the DeveloperHubServiceClient.GitHubOAuthCallback
 // method.
 type DeveloperHubServiceClientGitHubOAuthCallbackOptions struct {
@@ -63,6 +111,15 @@ type DeveloperHubServiceClientGitHubOAuthOptions struct {
 // method.
 type DeveloperHubServiceClientListGitHubOAuthOptions struct {
 	// placeholder for future optional parameters
+}
+
+// GeneratePreviewArtifactsResponse - Dockerfile and manifest artifacts generated as a preview are returned as a map
+type GeneratePreviewArtifactsResponse struct {
+	// An example generated file
+	PathToFile1WithExtension *string `json:"path/to/file1.withExtension,omitempty"`
+
+	// A second example file
+	PathToFile2WithExtension *string `json:"path/to/file2.withExtension,omitempty"`
 }
 
 // GitHubOAuthCallRequest - GitHubOAuth request object
@@ -118,9 +175,6 @@ type GitHubWorkflowProfile struct {
 	// The Azure Kubernetes Cluster Resource the application will be deployed to.
 	AksResourceID *string `json:"aksResourceId,omitempty"`
 
-	// Determines the type of manifests within the repository.
-	AuthStatus *ManifestType `json:"authStatus,omitempty"`
-
 	// Repository Branch Name
 	BranchName           *string               `json:"branchName,omitempty"`
 	DeploymentProperties *DeploymentProperties `json:"deploymentProperties,omitempty"`
@@ -143,6 +197,9 @@ type GitHubWorkflowProfile struct {
 
 	// Repository Owner
 	RepositoryOwner *string `json:"repositoryOwner,omitempty"`
+
+	// READ-ONLY; Determines the authorization status of requests.
+	AuthStatus *AuthorizationStatus `json:"authStatus,omitempty" azure:"ro"`
 
 	// READ-ONLY; The status of the Pull Request submitted against the users repository.
 	PrStatus *PullRequestStatus `json:"prStatus,omitempty" azure:"ro"`
@@ -283,13 +340,12 @@ type WorkflowClientGetOptions struct {
 	// placeholder for future optional parameters
 }
 
-// WorkflowClientListByResourceGroupOptions contains the optional parameters for the WorkflowClient.ListByResourceGroup method.
+// WorkflowClientListByResourceGroupOptions contains the optional parameters for the WorkflowClient.NewListByResourceGroupPager
+// method.
 type WorkflowClientListByResourceGroupOptions struct {
-	// The ManagedCluster resource associated with the workflows.
-	ManagedClusterResource *string
 }
 
-// WorkflowClientListOptions contains the optional parameters for the WorkflowClient.List method.
+// WorkflowClientListOptions contains the optional parameters for the WorkflowClient.NewListPager method.
 type WorkflowClientListOptions struct {
 	// placeholder for future optional parameters
 }
@@ -310,6 +366,9 @@ type WorkflowListResult struct {
 
 // WorkflowProperties - Workflow properties
 type WorkflowProperties struct {
+	// Properties used for generating artifacts such as Dockerfiles and manifests.
+	ArtifactGenerationProperties *ArtifactGenerationProperties `json:"artifactGenerationProperties,omitempty"`
+
 	// Profile of a github workflow.
 	GithubWorkflowProfile *GitHubWorkflowProfile `json:"githubWorkflowProfile,omitempty"`
 }