+++ sdk/resourcemanager/devhub/armdevhub/workflow_client.go
@@ -26,16 +26,18 @@ import (
 // WorkflowClient contains the methods for the Workflow group.
 // Don't use this type directly, use NewWorkflowClient() instead.
 type WorkflowClient struct {
-	host           string
-	subscriptionID string
-	pl             runtime.Pipeline
+	host                   string
+	subscriptionID         string
+	managedClusterResource *string
+	pl                     runtime.Pipeline
 }
 
 // NewWorkflowClient creates a new instance of WorkflowClient with the specified values.
-// subscriptionID - The ID of the target subscription.
-// credential - used to authorize requests. Usually a credential from azidentity.
-// options - pass nil to accept the default values.
-func NewWorkflowClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*WorkflowClient, error) {
+//   - subscriptionID - The ID of the target subscription.
+//   - managedClusterResource - The ManagedCluster resource associated with the workflows.
+//   - credential - used to authorize requests. Usually a credential from azidentity.
+//   - options - pass nil to accept the default values.
+func NewWorkflowClient(subscriptionID string, managedClusterResource *string, credential azcore.TokenCredential, options *arm.ClientOptions) (*WorkflowClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
@@ -48,19 +50,21 @@ func NewWorkflowClient(subscriptionID string, credential azcore.TokenCredential,
 		return nil, err
 	}
 	client := &WorkflowClient{
-		subscriptionID: subscriptionID,
-		host:           ep,
-		pl:             pl,
+		subscriptionID:         subscriptionID,
+		managedClusterResource: managedClusterResource,
+		host:                   ep,
+		pl:                     pl,
 	}
 	return client, nil
 }
 
 // CreateOrUpdate - Creates or updates a workflow
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-04-01-preview
-// resourceGroupName - The name of the resource group. The name is case insensitive.
-// workflowName - The name of the workflow resource.
-// options - WorkflowClientCreateOrUpdateOptions contains the optional parameters for the WorkflowClient.CreateOrUpdate method.
+//
+// Generated from API version 2022-09-16-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - workflowName - The name of the workflow resource.
+//   - options - WorkflowClientCreateOrUpdateOptions contains the optional parameters for the WorkflowClient.CreateOrUpdate method.
 func (client *WorkflowClient) CreateOrUpdate(ctx context.Context, resourceGroupName string, workflowName string, parameters Workflow, options *WorkflowClientCreateOrUpdateOptions) (WorkflowClientCreateOrUpdateResponse, error) {
 	req, err := client.createOrUpdateCreateRequest(ctx, resourceGroupName, workflowName, parameters, options)
 	if err != nil {
@@ -96,7 +100,7 @@ func (client *WorkflowClient) createOrUpdateCreateRequest(ctx context.Context, r
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01-preview")
+	reqQP.Set("api-version", "2022-09-16-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, parameters)
@@ -113,10 +117,11 @@ func (client *WorkflowClient) createOrUpdateHandleResponse(resp *http.Response) 
 
 // Delete - Deletes a workflow
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-04-01-preview
-// resourceGroupName - The name of the resource group. The name is case insensitive.
-// workflowName - The name of the workflow resource.
-// options - WorkflowClientDeleteOptions contains the optional parameters for the WorkflowClient.Delete method.
+//
+// Generated from API version 2022-09-16-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - workflowName - The name of the workflow resource.
+//   - options - WorkflowClientDeleteOptions contains the optional parameters for the WorkflowClient.Delete method.
 func (client *WorkflowClient) Delete(ctx context.Context, resourceGroupName string, workflowName string, options *WorkflowClientDeleteOptions) (WorkflowClientDeleteResponse, error) {
 	req, err := client.deleteCreateRequest(ctx, resourceGroupName, workflowName, options)
 	if err != nil {
@@ -152,7 +157,7 @@ func (client *WorkflowClient) deleteCreateRequest(ctx context.Context, resourceG
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01-preview")
+	reqQP.Set("api-version", "2022-09-16-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -169,10 +174,11 @@ func (client *WorkflowClient) deleteHandleResponse(resp *http.Response) (Workflo
 
 // Get - Gets a workflow.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-04-01-preview
-// resourceGroupName - The name of the resource group. The name is case insensitive.
-// workflowName - The name of the workflow resource.
-// options - WorkflowClientGetOptions contains the optional parameters for the WorkflowClient.Get method.
+//
+// Generated from API version 2022-09-16-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - workflowName - The name of the workflow resource.
+//   - options - WorkflowClientGetOptions contains the optional parameters for the WorkflowClient.Get method.
 func (client *WorkflowClient) Get(ctx context.Context, resourceGroupName string, workflowName string, options *WorkflowClientGetOptions) (WorkflowClientGetResponse, error) {
 	req, err := client.getCreateRequest(ctx, resourceGroupName, workflowName, options)
 	if err != nil {
@@ -208,7 +214,7 @@ func (client *WorkflowClient) getCreateRequest(ctx context.Context, resourceGrou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01-preview")
+	reqQP.Set("api-version", "2022-09-16-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -224,8 +230,9 @@ func (client *WorkflowClient) getHandleResponse(resp *http.Response) (WorkflowCl
 }
 
 // NewListPager - Gets a list of workflows associated with the specified subscription.
-// Generated from API version 2022-04-01-preview
-// options - WorkflowClientListOptions contains the optional parameters for the WorkflowClient.List method.
+//
+// Generated from API version 2022-09-16-preview
+//   - options - WorkflowClientListOptions contains the optional parameters for the WorkflowClient.NewListPager method.
 func (client *WorkflowClient) NewListPager(options *WorkflowClientListOptions) *runtime.Pager[WorkflowClientListResponse] {
 	return runtime.NewPager(runtime.PagingHandler[WorkflowClientListResponse]{
 		More: func(page WorkflowClientListResponse) bool {
@@ -266,7 +273,7 @@ func (client *WorkflowClient) listCreateRequest(ctx context.Context, options *Wo
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01-preview")
+	reqQP.Set("api-version", "2022-09-16-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -282,10 +289,11 @@ func (client *WorkflowClient) listHandleResponse(resp *http.Response) (WorkflowC
 }
 
 // NewListByResourceGroupPager - Gets a list of workflows within a resource group.
-// Generated from API version 2022-04-01-preview
-// resourceGroupName - The name of the resource group. The name is case insensitive.
-// options - WorkflowClientListByResourceGroupOptions contains the optional parameters for the WorkflowClient.ListByResourceGroup
-// method.
+//
+// Generated from API version 2022-09-16-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - options - WorkflowClientListByResourceGroupOptions contains the optional parameters for the WorkflowClient.NewListByResourceGroupPager
+//     method.
 func (client *WorkflowClient) NewListByResourceGroupPager(resourceGroupName string, options *WorkflowClientListByResourceGroupOptions) *runtime.Pager[WorkflowClientListByResourceGroupResponse] {
 	return runtime.NewPager(runtime.PagingHandler[WorkflowClientListByResourceGroupResponse]{
 		More: func(page WorkflowClientListByResourceGroupResponse) bool {
@@ -330,9 +338,9 @@ func (client *WorkflowClient) listByResourceGroupCreateRequest(ctx context.Conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01-preview")
-	if options != nil && options.ManagedClusterResource != nil {
-		reqQP.Set("managedClusterResource", *options.ManagedClusterResource)
+	reqQP.Set("api-version", "2022-09-16-preview")
+	if client.managedClusterResource != nil {
+		reqQP.Set("managedClusterResource", *client.managedClusterResource)
 	}
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
@@ -350,11 +358,12 @@ func (client *WorkflowClient) listByResourceGroupHandleResponse(resp *http.Respo
 
 // UpdateTags - Updates tags on a workflow.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2022-04-01-preview
-// resourceGroupName - The name of the resource group. The name is case insensitive.
-// workflowName - The name of the workflow resource.
-// parameters - Parameters supplied to the Update Workflow Tags operation.
-// options - WorkflowClientUpdateTagsOptions contains the optional parameters for the WorkflowClient.UpdateTags method.
+//
+// Generated from API version 2022-09-16-preview
+//   - resourceGroupName - The name of the resource group. The name is case insensitive.
+//   - workflowName - The name of the workflow resource.
+//   - parameters - Parameters supplied to the Update Workflow Tags operation.
+//   - options - WorkflowClientUpdateTagsOptions contains the optional parameters for the WorkflowClient.UpdateTags method.
 func (client *WorkflowClient) UpdateTags(ctx context.Context, resourceGroupName string, workflowName string, parameters TagsObject, options *WorkflowClientUpdateTagsOptions) (WorkflowClientUpdateTagsResponse, error) {
 	req, err := client.updateTagsCreateRequest(ctx, resourceGroupName, workflowName, parameters, options)
 	if err != nil {
@@ -390,7 +399,7 @@ func (client *WorkflowClient) updateTagsCreateRequest(ctx context.Context, resou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-04-01-preview")
+	reqQP.Set("api-version", "2022-09-16-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, parameters)